# cmd: improve performance of table rendering