```
name      old time/op    new time/op    delta 
String-4    18.5ms ± 4%    1.5ms ±16%  -92.07%  (p=0.004 n=5+6)

name      old alloc/op   new alloc/op   delta 
String-4    60.2MB ± 0%    0.5MB ± 0%  -99.18%  (p=0.004 n=5+6)

name      old allocs/op  new allocs/op  delta 
String-4     34.9k ± 0%    15.5k ± 0%  -55.67%  (p=0.002 n=6+6)
```